room package

+++ app/Http/Controllers/Dashboard/Room/AdminBackgroundRequestController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\Dashboard\Room;
 
 use App\Http\Controllers\Controller;
 use App\Http\Resources\Dashboard\Room\AdminBackgroundRequestsResource;
+use App\Support\PackageHelper;
 use Utd\Room\Entities\RequestBackgroundImage;
 use Illuminate\Http\Request;
 
@@ -12,18 +13,27 @@ class AdminBackgroundRequestController extends Controller
 
     public function index()
     {
+        if (!PackageHelper::isInstalled('room')) {
+            return collect();
+        }
         $data = RequestBackgroundImage::orderBy('id','desc')->get();
         return AdminBackgroundRequestsResource::collection($data);
     }
 
     public function show(string $id)
     {
+        if (!PackageHelper::isInstalled('room')) {
+            return null;
+        }
         $data = RequestBackgroundImage::find($id);
         return $data;
     }
 
     public function update(Request $request, string $id)
     {
+        if (!PackageHelper::isInstalled('room')) {
+            return 503;
+        }
         $item = RequestBackgroundImage::find($id);
         $item->status = $request->status;
         $item->update();
@@ -32,6 +42,9 @@ class AdminBackgroundRequestController extends Controller
 
     public function destroy(string $id)
     {
+        if (!PackageHelper::isInstalled('room')) {
+            return 503;
+        }
         $item = RequestBackgroundImage::find($id);
         $this->delete_img($item->img);
         $item->delete();
